Tweak parsing of ”Time Since Last Modified” (#37)
This is a proposed update to how we calculate time between a page was delivered (if that information exists) and when it was last modified. The PR introduces three new tests, one of them failing without the PR like this:

  1) headers #getTimeSinceLastModified should calculate diff between last-modified and date headers:

      AssertionError: expected 182735675.809 to equal 60
      + expected - actual

      -182735675.809
      +60

Parsing of date values from the headers is not changed. According to the specs, the format for the http Last-Modified and Date headers is not identical to that of javascript Date. But that’s left for a later patch.